requirements.txt
fastapi
python-multipart
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
psycopg2-binary
alembic